trim tests
