+++ tests/unit_tests/trim_tests.py
@@ -1,13 +1,6 @@
 import unittest
 from btb_tests import BtbTests
 
-import Bio
-from Bio import SeqIO
-from Bio.Seq import Seq
-# from Bio.Alphabet import generic_dna
-from Bio.SeqRecord import SeqRecord
-
-
 class TrimTests(BtbTests):
     adapter_path = './references/adapter.fasta'
 
@@ -34,28 +27,6 @@ class TrimTests(BtbTests):
         # Failure Case: adapter sequence file not found
         self.assertBashScript(1, ['./bin/trim.bash', './__does__/not/exist/', reads[0], reads[1], outputs[0], outputs[1]])
 
-    def write_fastq(self, filename, seq_records):
-        """
-            Write a fastq file to the temporary directory that tests run in
-
-            filename:  (str) name of output fastq file
-            sequences: (list/str) a list of strings that represent each sequence. Can also provide just a string if there is only one sequence
-            quality:   (int) uniform phred33 quality score for each base
-        """
-        # Handle a single sequence passed in directly as a string
-
-        # Write
-        with open(filename, "w") as file:
-            Bio.SeqIO.write(seq_records, file, "fastq")
-
-    def seq_record(self, seq_str, quality=93):
-        """
-            Makes a new Seq Record with uniform quality from a string of ATCG's 
-        """
-        seq = SeqRecord(Seq(seq_str))
-        seq.letter_annotations["phred_quality"] = [quality]*len(seq)
-        return seq
-
     def test_trimmomatric_trims(self):
         """
             Asserts that low quality reads are removed using a high quality and low quality examples
@@ -80,10 +51,10 @@ class TrimTests(BtbTests):
         self.assertBashScript(0, ['./bin/trim.bash', self.adapter_path, read_1, read_2, output_1, output_2])
 
         # Assert Output
-        trimmed = str(SeqIO.read(output_1, "fastq").seq)
+        trimmed = self.read_fastq(output_1)
         self.assertEqual(trimmed, high_quality_sequence)
 
-        trimmed = str(SeqIO.read(output_2, "fastq").seq)
+        trimmed = self.read_fastq(output_2)
         self.assertEqual(trimmed, high_quality_sequence)
 
 if __name__ == '__main__':
